+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * Added function `try_silent()` which tries to execute an expression and returns a string with the error message if the execution failed.
 
+* Added R6 object `Index` that provides a simple indexing interface for list elements.
+
 # oeli 0.1.0
 
 * Initial CRAN submission.